Bootstrapping InlineQueryResult classes #232
